[Denys Tykhonov] Refactored models structure
