+++ backend/src/model/__init__.py
@@ -1,0 +1,8 @@
+from src.model.course import Course
+from src.model.revoked_tokens import RevokedTokens
+from src.model.role import Role
+from src.model.subscription import Subscription
+from src.model.subscription_type import SubscriptionType
+from src.model.user import User
+from src.model.users_courses import UsersCourses
+from src.model.users_roles import UsersRoles
