+++ backend/src/model/revoked_tokens.py
@@ -1,0 +1,14 @@
+from app import db
+
+
+class RevokedTokens(db.Model):
+    id = db.Column(db.Integer, primary_key=True)
+    jti = db.Column(db.String(120))
+
+    def add(self):
+        db.session.add(self)
+        db.session.commit()
+
+    @classmethod
+    def is_jti_blacklisted(cls, jti):
+        return bool(cls.query.filter_by(jti=jti).first())
