+++ backend/src/model/users_courses.py
@@ -1,0 +1,7 @@
+from app import db
+
+
+class UsersCourses(db.Model):
+    id = db.Column(db.Integer(), primary_key=True)
+    user_id = db.Column(db.Integer(), db.ForeignKey('user.id', ondelete='CASCADE'))
+    course_id = db.Column(db.Integer(), db.ForeignKey('course.id', ondelete='CASCADE'))
